requirements.txt
asgiref==3.2.7
atomicwrites==1.4.0
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.6
django-cors-headers==3.3.0
djangorestframework==3.11.0
drf-yasg==1.17.1
idna==2.9
importlib-metadata==1.6.0
inflection==0.4.0
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
more-itertools==8.3.0
packaging==20.4
pluggy==0.13.1
psycopg2==2.8.5
py==1.8.1
pyDataverse @ git+https://gitlab.whiteaster.com/public_data/pydataverse-upgrade.git@92e0f0c4d6d45025d2871ea7a315a2477e4331b6
pyparsing==2.4.7
pysolr==3.9.0
pytest==5.4.2
pytz==2020.1
redis==3.5.3
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.15.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.9
wcwidth==0.1.9
zipp==3.1.0
